чутка поменял структуру
